@@ -35,32 +35,33 @@
     ]
   },
   "dependencies": {
-    "@chakra-ui/icons": "^2.0.18",
-    "@chakra-ui/react": "^2.5.5",
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
+    "@chakra-ui/icons": "^2.0.19",
+    "@chakra-ui/next-js": "^2.1.3",
+    "@chakra-ui/react": "^2.6.1",
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
     "@fontsource/questrial": "^4.5.11",
-    "framer-motion": "^10.11.2",
-    "next": "^13.3.0",
+    "framer-motion": "^10.12.8",
+    "next": "^13.4.1",
     "next-pwa": "^5.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
-    "sass": "^1.61.0"
+    "sass": "^1.62.1"
   },
   "devDependencies": {
-    "@next/eslint-plugin-next": "^13.3.0",
+    "@next/eslint-plugin-next": "^13.4.1",
     "@types/gtag.js": "0.0.12",
-    "@types/node": "^18.15.11",
-    "@types/react": "^18.0.33",
-    "@typescript-eslint/eslint-plugin": "^5.57.1",
-    "@typescript-eslint/parser": "^5.57.1",
-    "eslint": "^8.37.0",
-    "eslint-config-next": "^13.3.0",
+    "@types/node": "^20.1.0",
+    "@types/react": "^18.2.6",
+    "@typescript-eslint/eslint-plugin": "^5.59.5",
+    "@typescript-eslint/parser": "^5.59.5",
+    "eslint": "^8.40.0",
+    "eslint-config-next": "^13.4.1",
     "eslint-plugin-react": "^7.32.2",
     "husky": "^8.0.3",
-    "lint-staged": "^13.2.0",
-    "prettier": "^2.8.7",
-    "typescript": "^5.0.3"
+    "lint-staged": "^13.2.2",
+    "prettier": "^2.8.8",
+    "typescript": "^5.0.4"
   }
 }
